blacksmith/experiments/torch/BOUNTIES/falcon3-1b/requirements.txt
# SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
#
# SPDX-License-Identifier: Apache-2.0
#
# Dependencies for Falcon3-1B LoRA Training
# ==========================================

# Core PyTorch
torch>=2.1.0

# Transformers and PEFT for LoRA
transformers>=4.36.0
peft>=0.7.0

# Datasets
datasets>=2.16.0

# Tokenizers
tokenizers>=0.15.0

# Logging and tracking
wandb>=0.16.0

# Plotting
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Progress bars
tqdm>=4.66.0

# Numerical operations
numpy>=1.24.0

# For TT-N150 support (install separately from Tenstorrent)
# torch-xla (TT variant)

